implement when click warning item.

+++ MonitoringTourSystem/MonitoringTourSystem/Models/Warning.cs
@@ -8,11 +8,13 @@ namespace MonitoringTourSystem.Models
     public class Warning
     {
         public string CategoryWarnig { get; set; }
-        public int Distance { get; set; }
+        public double Distance { get; set; }
         public double Lat { get; set; }
         public double Long { get; set; }
         public string WarningName { get; set; }
         public string DescriptionWarning { get; set; }
+
+
 
     }
 }
